refactor(test-util): moved `marbles.spec-util` to common `util` folder
