+++ packages/core/src/effects/effects.combiner.spec.ts
@@ -1,6 +1,6 @@
 import { mapTo, tap } from 'rxjs/operators';
+import { Marbles } from '../../../util/marbles.spec-util';
 import { HttpRequest, HttpResponse } from '../http.interface';
-import { Marbles } from '../util/marbles.spec-util';
 import { combineEffects, combineMiddlewareEffects, combineRoutes } from './effects.combiner';
 import { Effect, GroupedEffects } from './effects.interface';
